requirements.txt
streamlit
torch
torchvision
pillow
pandas
numpy
scikit-learn
